# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Freshness Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PageFreshnessTriageStatus =
  (typeof PageFreshnessTriageStatus)[keyof typeof PageFreshnessTriageStatus];

export const PageFreshnessTriageStatus = {
  critical: "critical",
  review: "review",
  healthy: "healthy",
} as const;

export type PageFreshnessTrafficTrend =
  (typeof PageFreshnessTrafficTrend)[keyof typeof PageFreshnessTrafficTrend];

export const PageFreshnessTrafficTrend = {
  up: "up",
  down: "down",
  stable: "stable",
} as const;

export interface PageFreshness {
  id: number;
  url: string;
  title?: string;
  lastUpdated: string;
  clicks30d: number;
  clicksPrev30d: number;
  wordCount: number;
  excerpt?: string;
  /** 0-100 score indicating content freshness */
  freshnessScore: number;
  /** 0-100 score indicating content decay (higher = more decayed) */
  decayScore: number;
  triageStatus: PageFreshnessTriageStatus;
  /** Whether the page is likely to be cited by AI search */
  aiCitationLikely: boolean;
  aiCitationReason?: string;
  daysSinceUpdate: number;
  trafficTrend: PageFreshnessTrafficTrend;
  createdAt: string;
}

export interface CreatePageInput {
  url: string;
  title?: string;
  lastUpdated: string;
  clicks30d: number;
  clicksPrev30d: number;
  wordCount: number;
  excerpt?: string;
}

export interface CsvUploadInput {
  /** CSV content as a string */
  csvData: string;
}

export interface CsvUploadResult {
  imported: number;
  skipped: number;
  errors: string[];
}

export interface DashboardStats {
  totalPages: number;
  /** Percentage of pages under 90 days old */
  freshPercent: number;
  criticalCount: number;
  reviewCount: number;
  healthyCount: number;
  avgFreshnessScore: number;
  aiCitationReadyCount: number;
}