# lib/api-zod/src/generated/types/createPageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Freshness Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePageInput {
  url: string;
  title?: string;
  lastUpdated: Date;
  clicks30d: number;
  clicksPrev30d: number;
  wordCount: number;
  excerpt?: string;
}