# lib/api-zod/src/generated/types/csvUploadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Freshness Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CsvUploadInput {
  /** CSV content as a string */
  csvData: string;
}