# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Freshness Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalPages: number;
  /** Percentage of pages under 90 days old */
  freshPercent: number;
  criticalCount: number;
  reviewCount: number;
  healthyCount: number;
  avgFreshnessScore: number;
  aiCitationReadyCount: number;
}