# lib/api-zod/src/generated/types/pageFreshness.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Freshness Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PageFreshnessTrafficTrend } from "./pageFreshnessTrafficTrend";
import type { PageFreshnessTriageStatus } from "./pageFreshnessTriageStatus";

export interface PageFreshness {
  id: number;
  url: string;
  title?: string;
  lastUpdated: Date;
  clicks30d: number;
  clicksPrev30d: number;
  wordCount: number;
  excerpt?: string;
  /** 0-100 score indicating content freshness */
  freshnessScore: number;
  /** 0-100 score indicating content decay (higher = more decayed) */
  decayScore: number;
  triageStatus: PageFreshnessTriageStatus;
  /** Whether the page is likely to be cited by AI search */
  aiCitationLikely: boolean;
  aiCitationReason?: string;
  daysSinceUpdate: number;
  trafficTrend: PageFreshnessTrafficTrend;
  createdAt: Date;
}